refactor(recording): rename accessibility permission registrar
