+++ apps/whispering/src/routes/(app)/_layout-utils/register-accessibility-permission.ts
